+++ test/fixtures/script.js
@@ -1,0 +1,8 @@
+
+var oldlib = require('./old-lib')
+
+run()
+
+function run() {
+  oldlib.old()
+}